Added the ALU
The computer now has the ability to add and subtract the numbers in the A and B registers.

The 8-bit adder circuit that was used in the ALU is optimized specifically for Minecraft, and uses a method similar to a Carry-Lookahead adder. I got it from a YouTube video by "mattbatwings" with the title "Synced Addition."